% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pkgrefs.R
\name{as_pkgrefs}
\alias{as_pkgrefs}
\alias{as_pkgrefs.default}
\alias{as_pkgrefs.sessionInfo}
\title{Convert Data Structures into Package References}
\usage{
as_pkgrefs(x, ...)

\method{as_pkgrefs}{default}(x, ...)

\method{as_pkgrefs}{sessionInfo}(x, ...)
}
\arguments{
\item{x, }{currently supported data structure(s) are: output from \code{\link[=sessionInfo]{sessionInfo()}}, a character vector of package names}

\item{..., }{not used}
}
\value{
a vector of package references
}
\description{
This generic function converts several standard data structures into a vector of package references, which in turn
can be used as the first argument of the function \code{\link[=resolve]{resolve()}}. This function guessimates the possible sources of the
packages. But we strongly recommend manually reviewing the detected packages before using them for \code{\link[=resolve]{resolve()}}.
}
\examples{
as_pkgrefs(sessionInfo())
if (interactive()) {
   require(rang)
   require(pkgsearch)
   graph <- resolve(as_pkgrefs(sessionInfo()))
}
}
