% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rang.R
\name{create_turing}
\alias{create_turing}
\title{Create executable research compendium according to the Turing Way}
\usage{
create_turing(path, add_rang = TRUE, ...)
}
\arguments{
\item{path}{character, path to the project root}

\item{add_rang}{logical, whether to run \code{\link[=use_rang]{use_rang()}} to \code{path}}

\item{...}{additional parameters pass to \code{\link[=use_rang]{use_rang()}}}
}
\value{
path, invisibly
}
\description{
This \code{usethis}-style function creates an executable research compendium according to the Turing Way.
}
\details{
A research compendium according to the Turing Way contains:
\itemize{
\item \code{data_raw}: a directory to hold the raw data
\item \code{data_clean}: a directory to hold the processed data
\item \code{code}: a directory to hold computer code
\item \code{CITATION}: a file holding citation information
\item \code{paper.Rmd}: a manuscript
The structure is just a suggestion. For example, the manuscript can be in other format.
}

According to the Turing Way, an executable research compendium should
\enumerate{
\item Files should be organized in a conventional folder structure;
\item Data, methods, and output should be clearly separated;
\item The computational environment should be specified.
}

With \code{add_rang}, the computational environment can be recorded and reconstructed later.
}
\references{
\href{https://the-turing-way.netlify.app/reproducible-research/compendia.html}{The Turing Way: Research Compendia}
Gorman, KB, Williams TD. and Fraser WR (2014). \href{http://dx.doi.org/10.1371/journal.pone.0090081}{Ecological Sexual Dimorphism and Environmental Variability within a Community of Antarctic Penguins (Genus Pygoscelis)}. PLoS ONE 9(3):e90081.
}
\seealso{
\code{\link[=use_rang]{use_rang()}}
}
