% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{dockerize}
\alias{dockerize}
\alias{dockerize_granlist}
\alias{dockerise}
\alias{dockerise_granlist}
\title{Dockerize The Resolved Result}
\usage{
dockerize(
  granlist,
  output_dir,
  image = c("r-ver", "rstudio", "tidyverse", "verse", "geospatial"),
  granlist_as_comment = TRUE
)

dockerize_granlist(...)

dockerise(...)

dockerise_granlist(...)
}
\arguments{
\item{granlist}{output from \code{\link[=resolve]{resolve()}}}

\item{output_dir}{where to put the Docker file}

\item{image}{character, which versioned Rocker image to use. Can only be "r-ver", "rstudio", "tidyverse", "verse", "geospatial"}

\item{granlist_as_comment}{logical, whether to write granlist and the steps to reproduce
the file to \code{path} as comment}

\item{...}{arguments to be passed to \code{dockerize}}
}
\value{
\code{output_dir}, invisibly
}
\description{
This function exports the result from \code{\link[=resolve]{resolve()}} to a Docker file.
}
\examples{
\donttest{
if (interactive()) {
    graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                    snapshot_date = "2020-01-16")
    dockerize(graph, ".")
}
}
}
\seealso{
\code{\link[=resolve]{resolve()}}, \code{\link[=export_granlist]{export_granlist()}}
}
