% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{export_granlist}
\alias{export_granlist}
\title{Export The Resolved Result As Installation Script}
\usage{
export_granlist(
  granlist,
  path,
  granlist_as_comment = TRUE,
  verbose = TRUE,
  lib = NA
)
}
\arguments{
\item{granlist}{output from \code{\link[=resolve]{resolve()}}}

\item{path}{character, path of the exported installation script}

\item{granlist_as_comment}{logical, whether to write granlist and the steps to reproduce
the file to \code{path} as comment}

\item{verbose}{logical, pass to \code{\link[=install.packages]{install.packages()}}, the negated value is also passed as \code{quiet} to both \code{\link[=install.packages]{install.packages()}}
and \code{\link[=download.file]{download.file()}}.}

\item{lib}{character, pass to \code{\link[=install.packages]{install.packages()}}. By default, it is NA (to install the packages to the default location)}
}
\value{
\code{path}, invisibly
}
\description{
This function exports the results from \code{\link[=resolve]{resolve()}} to an installation script that can be run in a fresh R environment.
}
\examples{
\donttest{
if (interactive()) {
    graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                    snapshot_date = "2020-01-16")
    export_granlist(graph, "gran.R")
}
}
}
