% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{query_sysreqs}
\alias{query_sysreqs}
\title{Query for System Requirements}
\usage{
query_sysreqs(rang, os = "ubuntu-20.04")
}
\arguments{
\item{rang}{output from \code{\link[=resolve]{resolve()}}}

\item{os}{character, which OS to query for system requirements}
}
\value{
S3 object \code{rang}
}
\description{
This function takes an S3 object returned from \code{\link[=resolve]{resolve()}} and (re)queries the System Requirements.
}
\examples{
\donttest{
if (interactive()) {
    graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                snapshot_date = "2020-01-16", get_sysreqs = FALSE)
    graph$deps_sysreqs
    graph2 <- query_sysreqs(graph, os = "ubuntu-20.04")
    graph2$deps_sysreqs
}
}
}
\seealso{
\code{\link[=resolve]{resolve()}}
}
