% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve.R
\name{resolve}
\alias{resolve}
\title{Resolve Dependencies Of R Packages}
\usage{
resolve(
  pkgs,
  snapshot_date,
  no_enhances = TRUE,
  no_suggests = TRUE,
  get_sysreqs = TRUE,
  os = "ubuntu-20.04",
  verbose = FALSE
)
}
\arguments{
\item{pkgs}{character vector of R packages to resolve}

\item{snapshot_date}{Snapshot date, if not specified, assume to be a month ago}

\item{no_enhances}{logical, whether to ignore packages in the "Enhances" field}

\item{no_suggests}{logical, whether to ignore packages in the "Suggests" field}

\item{get_sysreqs}{logical, whether to query for System Requirements}

\item{os}{character, which OS to query for system requirements}

\item{verbose}{logical, whether to display messages}
}
\value{
S3 object \code{granlist}
}
\description{
This function recursively queries dependencies of R packages at a specific snapshot time. The dependency graph can then be used to recreate the computing environment. The data on dependencies are provided by R-hub.
}
\examples{
\donttest{
if (interactive()) {
    graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                snapshot_date = "2020-01-16")
    graph
}
}
}
\seealso{
\code{\link[=dockerize]{dockerize()}}
}
