% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{dockerize}
\alias{dockerize}
\title{Docker The Resolved Result}
\usage{
dockerize(
  granlist,
  output_dir,
  image = c("r-ver", "rstudio", "tidyverse", "verse", "geospatial")
)
}
\arguments{
\item{granlist}{output from \code{\link[=resolve]{resolve()}}}

\item{output_dir}{where to put the Docker file}

\item{image}{character, which versioned Rocker image to use. Can only be "r-ver", "rstudio", "tidyverse", "verse", "geospatial"}
}
\value{
output_dir, invisibly
}
\description{
This function exports the result from \code{\link[=resolve]{resolve()}} to a Docker file.
}
\examples{
\donttest{
if (interactive()) {
    graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                    snapshot_date = "2020-01-16")
    dockerize(graph, ".")
}
}
}
