% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{dockerize}
\alias{dockerize}
\title{Docker The Resolved Result}
\usage{
dockerize(granlist, output_dir)
}
\arguments{
\item{granlist}{output from \code{\link[=resolve]{resolve()}}}

\item{output_dir}{where to put the Docker file}
}
\value{
output_dir, invisibly
}
\description{
This function exports the result from \code{\link[=resolve]{resolve()}} to a Docker file.
}
\examples{
\donttest{
graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                snapshot_date = "2020-01-16")
}
}
