% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{dockerize}
\alias{dockerize}
\alias{dockerize_granlist}
\alias{dockerise}
\alias{dockerise_granlist}
\title{Dockerize The Resolved Result}
\usage{
dockerize(
  granlist,
  output_dir,
  image = c("r-ver", "rstudio", "tidyverse", "verse", "geospatial"),
  granlist_as_comment = TRUE,
  verbose = TRUE,
  lib = NA,
  cran_mirror = "https://cran.r-project.org/",
  check_cran_mirror = TRUE
)

dockerize_granlist(...)

dockerise(...)

dockerise_granlist(...)
}
\arguments{
\item{granlist}{output from \code{\link[=resolve]{resolve()}}}

\item{output_dir}{where to put the Docker file}

\item{image}{character, which versioned Rocker image to use. Can only be "r-ver", "rstudio", "tidyverse", "verse", "geospatial".
This applies only when R version <= 3.1.}

\item{granlist_as_comment}{logical, whether to write granlist and the steps to reproduce
the file to \code{path} as comment}

\item{verbose}{logical, pass to \code{\link[=install.packages]{install.packages()}}, the negated value is also passed as \code{quiet} to both \code{\link[=install.packages]{install.packages()}}
and \code{\link[=download.file]{download.file()}}.}

\item{lib}{character, pass to \code{\link[=install.packages]{install.packages()}}. By default, it is NA (to install the packages to the default location)}

\item{cran_mirror}{character, which CRAN mirror to use}

\item{check_cran_mirror}{logical, whether to check the CRAN mirror}

\item{...}{arguments to be passed to \code{dockerize}}
}
\value{
\code{output_dir}, invisibly
}
\description{
This function exports the result from \code{\link[=resolve]{resolve()}} to a Docker file.
}
\details{
The idea behind this is to determine the installation order of R packages locally. Then, the installation script can be depolyed to another
fresh R session to install R packages. \code{\link[=dockerize]{dockerize()}} is a more reasonable way because a fresh R session with all system requirements
is provided. The current approach does not work in R < 2.1.0.
}
\examples{
\donttest{
if (interactive()) {
    graph <- resolve(pkgs = c("openNLP", "LDAvis", "topicmodels", "quanteda"),
                    snapshot_date = "2020-01-16")
    dockerize(graph, ".")
}
}
}
\references{
Ripley, B. (2005) \href{https://cran.r-project.org/doc/Rnews/Rnews_2005-1.pdf}{Packages and their Management in R 2.1.0.} R News, 5(1):8--11.
}
\seealso{
\code{\link[=resolve]{resolve()}}, \code{\link[=export_granlist]{export_granlist()}}
}
